container is running but not make docker image
I manually installed Amber Suit on an Ubuntu container without a Docker file. As a result, all installations were successful and testing was done. However, when I tried to create an image from the running container, it was not successful due to the issue of no space left on the device. Even after increasing the space, the issue could not be fixed.